버그 수정.

+++ Common/Network/NetUtil.h
@@ -2,6 +2,6 @@
 
 class NetUtil {
 	static bool HostNameToIp(const char* _host, OUT std::string& _ip);
-	static uint16_t HostNameToIp(const char* _host);
+	static int32_t HostNameToIp(const char* _host);
 	static bool IsUsedAddr(sockaddr_in _addr, bool _b_tcp);
 };
